pipeline/monitoring_prediction/requirements.txt
flask
protobuf==3.19.6
tensorflow-data-validation==1.12.0
tensorflow-metadata==1.12.0
google-cloud-storage==1.44.0
fsspec==2023.1.0
gcsfs==2023.1.0
gspread==5.7.2
ipython==7.34.0